requirements.txt
youtube-transcript-api==0.6.2
pandas==2.2.3

langchain==0.3.3
langchain-community==0.3.2
langchain-core==0.3.12

langchain-google-genai==2.0.1

chromadb==0.5.5

python-dotenv==1.0.1
